fix(emoji-size-in-chat): Merged all Emoji functions into EmojiUtils
