--- src/libs/Emoji/trimEmojiUnicode.js
@@ -1,8 +1,0 @@
-/**
- * Function to remove Skin Tone and utf16 surrogates from Emoji
- * @param {String} emojiCode
- * @returns {String}
- */
-export default function trimEmojiUnicode(emojiCode) {
-    return emojiCode.replace(/(fe0f|1f3fb|1f3fc|1f3fd|1f3fe|1f3ff)$/, '').trim();
-}
